LPS-161981 batch-engine - convert to builder pattern
